refactor(cockpit): split background session routes
Refs AC-778, AC-779.
